feat(suite-native): Refactor feedback-form package structure
