+++ suite-native/feedback-form/src/components/EmojiRatingButton.tsx
