+++ suite-native/feedback-form/src/components/FeedbackCard.tsx
